requirements.txt
opencv-python==4.8.0.74
numpy==1.24.3
pytesseract==0.3.10
Pillow==10.0.0
neo4j==5.11.0
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.0.0
flask==2.3.3 